LWF Changes

+++ myDatasets.py
@@ -190,7 +190,7 @@ class ConcatDataset(torch.utils.data.Dataset):
                 for t in dataset.labels:
                     self.targets.append(t)
 
-        self.data = np.array(self.data)
+        self.data = np.array(self.data, dtype=np.float32)
         self.targets = np.array(self.targets)
 
         self.data = torch.Tensor(self.data)
